:recycle: Refactor ERD commands into a dedicated module
Moved the ERD-related logic to a new `erdCommand` module for better organization and separation of concerns. This refactoring includes:

- Renaming and relocating `buildCommand` and related files under `src/cli/erdCommand`.
- Adding a dedicated `erdCommand` definition in `index.ts` and connecting it to the main program.
- Updated tests to reflect the new structure, ensuring all commands and options function as expected.

This change improves maintainability and prepares the CLI for future expansions.
