--- frontend/packages/cli/src/cli/commands/index.ts
@@ -1,2 +1,0 @@
-import { buildCommand } from './buildCommand.js'
-export { buildCommand }
