+++ frontend/packages/cli/src/cli/erdCommand/getInputContent.test.ts
