+++ frontend/packages/cli/src/cli/erdCommand/getInputContent.ts
